+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 * integrate a finite state machine to control valid state transitions
 * support for both automatic and conditional (success-based) state transitions
 
+* add ng packages as namespace in dirs 'unit' and 'logprep' as preparation for new architecture implementation
+* add abstract EventMetadata class and KafkaInputMetadata class
+
 ### Bugfix
 
 ## 17.0.0